% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listIso.R
\docType{data}
\name{listIso}
\alias{listIso}
\title{Isotopic abundances of atoms typically found in nucleic acids.}
\format{
A list of 7:
\describe{
     \item{H}{H isotopic abundances}
     \item{C}{C isotopic abundances}
     \item{N}{N isotopic abundances}
     \item{O}{O isotopic abundances}
     \item{D}{D isotopic abundances. Same as H; placeholder for non-natural isotopic compositions}
     \item{P}{P isotopic abundances}
     \item{K}{K isotopic abundances}
   }
}
\source{
\url{https://www.ciaaw.org/isotopic-abundances.htm/}
}
\usage{
listIso
}
\description{
A dataset containing the isotopic abundances of H, C, N, O, D, P, and K.
}
\keyword{datasets}
