% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massR.R
\name{massR}
\alias{massR}
\title{Oligonucleotide molecular weight and m/z calculator}
\usage{
massR(seq, DC, K41C = 6.730244)
}
\arguments{
\item{seq}{A list created with the \code{sequenceR} function.}

\item{DC}{A numeric value defining the percentage of in solution deuterium.}

\item{K41C}{A numeric value defining the percentage of 41K isotope. Defaults to the natural abundance of 6.730244%.}
}
\value{
A list with the monoisotopic molecular weight and m/z, and the monoisotopic molecular weight and m/z.
}
\description{
Oligonucleotide molecular weight and m/z calculator
}
\examples{
massR(DC = 9, K41C = 6.730244 seq = sequencer)
}
