% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligor.R
\name{oligor}
\alias{oligor}
\title{Software suite for oligonucleotide native MS data treatment}
\usage{
oligor()
}
\value{
Launches the shiny app.
}
\description{
Software suite for oligonucleotide native MS data treatment
}
\examples{
oligor()
}
