% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpositionR.R
\name{peak.positionR}
\alias{peak.positionR}
\title{Oligonucleotide isotopic peak distribution calculation by FFT based on J. Proteome Res. 2018 Jan 5; 17(1): 751–758}
\usage{
peak.positionR(nrPeaks.user, DC, K41C = 6.730244, seq, MonoMW)
}
\arguments{
\item{nrPeaks.user}{An integer controlling the number of calculated peaks.}

\item{DC}{A numeric value defining the percentage of in solution deuterium.}

\item{K41C}{A numeric value defining the percentage of 41K isotope. Defaults to the natural abundance of 6.730244%.}

\item{seq}{A list created with the \code{sequenceR} function.}

\item{MonoMW}{A numeric value, the monoisotopic mass of the oligonucleotide, which can be imported from the \code{massR} function.}
}
\value{
A dataframe containing the isotopic distribution.
}
\description{
Oligonucleotide isotopic peak distribution calculation by FFT based on J. Proteome Res. 2018 Jan 5; 17(1): 751–758
}
\examples{
peak.positionR(nrPeaks.user = 32, DC = 9, K41C = 6.730244 seq = sequencer, MonoMW = massr$MonoMW)
}
