% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpicking.R
\name{peakpicking}
\alias{peakpicking}
\title{Peak picking function based on the peakPick package}
\usage{
peakpicking(raw.data, neighlim = 5, deriv.lim = 10000, int.thresh = 0.02)
}
\arguments{
\item{raw.data}{A data frame with an m/z column called \code{mz} and an intensity column called \code{int}.}

\item{neighlim}{An integer limit for how far apart peaks must be. Peak pairs closer than or equal to neighlim to each other have the lesser peak eliminated.}

\item{deriv.lim}{A numeric upper limit for the estimated derivative for a point to be considered for a peak call}

\item{int.thresh}{A numeric intensity threshold value, relative to a maximum intensity of 1, to discard low intensity false positives.}
}
\value{
A dataframe with an additional column \code{peak} equal to 0 in absence of peak or the peak intensity if a peak is detected at that position
}
\description{
Peak picking function based on the peakPick package
}
\examples{
peakpicking(raw.data = a.data.frame, neighlim = 5, deriv.lim = 10000, int.thresh = 0.02)neighlim = 5, deriv.lim = 10000, int.thresh = 0.02
}
