% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceR.R
\name{sequenceR}
\alias{sequenceR}
\title{Oligonucleotide sequence analysis}
\usage{
sequenceR(z, K, sequence, nX.user.input, nX.select)
}
\arguments{
\item{z}{An integer, defining the absolute value of the charge state.}

\item{K}{An integer, the number of K+ adducts.}

\item{nX.user.input}{An integer, the number of exchangeable sites. Use \code{''} to use the presets instead.}

\item{nX.select}{A string defining the preset to use for the number of exchangeable sites: all sites = \code{'A'}, no phosphates = \code{'B'}, no phosphates nor termini = \code{'C'}}
}
\value{
A list containing the charge state \code{z}, the number of potassium adducts \code{K},
the total number of nucleotides \code{nb_nt},
the number of each nucleotide \code{nbA}, \code{nbC}, \code{nbG}, \code{nbT},
the number of each atom \code{nH}, \code{nC}, \code{nO}, \code{nN}, \code{nP}, \code{nK},
the number of exchangeable sites \code{nX},
the number of phosphates \code{nb_PO}, and the number of neutralized phosphates \code{nb_POH}
}
\description{
Oligonucleotide sequence analysis
}
\examples{
sequenceR(z=4, K=2, sequence="TTGGGTGGGTGGGTGGGT", nX.user.input='', nX.select='C')
}
